+++ src/components/survey/SurveyView/SurveyProgress/SurveyProgress.tsx
@@ -5,25 +5,7 @@ interface SurveyProgressProps {
   currentIndex: number;
   totalCount: number;
 }
-
-/*const useStyles = makeStyles((theme: Theme) =>
-  createStyles({
-    root: {
-      display: 'flex',
-      width: '100%',
-      height: 5,
-      alignItems: 'center',
-      justifyContent: 'center'
-    },
-    item: {
-      backgroundColor: '#ccc',
-      margin: 3,
-      height: "100%",
-      flexGrow: 1,
-      maxWidth: 40,
-    },
-  }),
-);*/
+const maxPageBarWidth = 45;
 
 const SurveyProgress: React.FC<SurveyProgressProps> = (props) => {
   return (
@@ -32,28 +14,30 @@ const SurveyProgress: React.FC<SurveyProgressProps> = (props) => {
       className="d-flex w-100 align-items-center justify-content-center"
       style={{ height: 5 }}
     >
-      {
-        Array.from(Array(props.totalCount).keys()).map(
-          index => (
-            <div
-              key={index.toString()}
-              className={
-                clsx(
-                  'h-100',
-                  'flex-grow-1',
-                  {
-                    'bg-grey-2': index > props.currentIndex,
-                    'bg-primary': index <= props.currentIndex,
-                  })}
-              style={{ margin: 3, maxWidth: 40 }}
-            >
-            </div>
+      <div className="d-flex justify-content-evenly h-100"
+        style={{ width: (maxPageBarWidth + 10) * props.totalCount, maxWidth: '100%' }}
+      >
+        {
+          Array.from(Array(props.totalCount).keys()).map(
+            index => (
+              <div
+                key={index.toString()}
+                className={
+                  clsx(
+                    'h-100',
+                    'flex-grow-1',
+                    {
+                      'bg-grey-2': index > props.currentIndex,
+                      'bg-primary': index <= props.currentIndex,
+                    })}
+                style={{ maxWidth: maxPageBarWidth }}
+              >
+              </div>
+            )
           )
-        )
-      }
+        }
+      </div>
     </div>
-
-
   );
 };